backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
cryptography==42.0.1

# WebSocket Support
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

# Celery & Background Tasks
celery==5.3.6
redis==5.0.1

# API Integration
requests==2.31.0
requests-oauthlib==1.3.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Analytics & Export
pandas==2.2.0
reportlab==4.0.9

# API Documentation
drf-spectacular==0.27.1

# Development
django-debug-toolbar==4.2.0
